+++ src/core/constants/paths.ts
@@ -11,16 +11,35 @@ export const PROJECT_PATH = path.join(__dirname, "../../..");
 const config = ConfigManager.readConfig();
 export const HOME_DIR = os.homedir();
 
-export const COM_MOJANG_PATH =
-    config.output !== "preview"
-        ? path.join(
-              HOME_DIR,
-              "AppData/Local/Packages/Microsoft.MinecraftUWP_8wekyb3d8bbwe/LocalState/games/com.mojang"
-          )
-        : path.join(
-              HOME_DIR,
-              "AppData/Local/Packages/Microsoft.MinecraftWindowsBeta_8wekyb3d8bbwe/LocalState/games/com.mojang"
-          );
+export const COM_MOJANG_PATH = ((): string => {
+    switch (config.output) {
+        case "stable":
+            return path.join(
+                HOME_DIR,
+                "AppData/Roaming/Minecraft Bedrock/Users/Shared/games/com.mojang"
+            );
+        case "preview":
+            return path.join(
+                HOME_DIR,
+                "AppData/Roaming/Minecraft Bedrock Preview/Users/Shared/games/com.mojang"
+            );
+        case "stable_uwp":
+            return path.join(
+                HOME_DIR,
+                "AppData/Local/Packages/Microsoft.MinecraftUWP_8wekyb3d8bbwe/LocalState/games/com.mojang"
+            );
+        case "preview_uwp":
+            return path.join(
+                HOME_DIR,
+                "AppData/Local/Packages/Microsoft.MinecraftWindowsBeta_8wekyb3d8bbwe/LocalState/games/com.mojang"
+            );
+        default:
+            return path.join(
+                HOME_DIR,
+                "AppData/Roaming/Minecraft Bedrock/Users/Shared/games/com.mojang"
+            );
+    }
+})();
 
 export const OUTPUT_BEHAVIOUR_PACK_PATH = path.join(
     COM_MOJANG_PATH,