+++ changelog.md
@@ -2,3 +2,9 @@
 - NPM Module:
 - Now supports preview packages, to enable make sure output is set to "preview"
 - Added support for @minecraft/server-graphics
+# 1.0.2
+- Schema now is hosted on github, so it works if multiple people work on an project (run "mc repair" to fix you schema)
+- Added "awaitWriteFinish" to config, use if you experience issues with the tool transfering data too fast
+- Updated output paths:
+- "stable" and "preview" now point to the gdk paths
+- "stable_uwp" and "preview_uwp" can be used if your still on uwp